### 2026-02-19 #10083 @editorialbot
[REVIEW]: PICNIC: an open-source Python library for preprocessing of dynamic Positron Emission Tomography (PET) brain imaging data
**Submitting author:** @FrancescaZanderigo (<a href="http://orcid.org/0000-0001-6510-0676">Francesca Zanderigo</a>)
**Repository:** https://github.com/bil-mind-nyspi/PICNIC
**Branch with paper.md** (empty if default branch): main
**Version:** v0.1.4
