### 2026-02-19 #10083 @FrancescaZanderigo
Please note that the correct repository is
*https://github.com/bil-mind-nyspi/PICNIC
<https://github.com/bil-mind-nyspi/PICNIC>*,
as we stated in the latest version of the paper.md that we updated on
September 30 2025 (here attached for reference).

My best,
Francesca

Francesca Zanderigo, PhD
Associate Professor, Columbia University, Department of Psychiatry
Co-Director of Brain Imaging, Molecular Imaging and Neuropathology Area,
New York State Psychiatric Institute
1051 Riverside Drive, New York, NY, 10032
***@***.*** ***@***.***>

On Thu, Feb 19, 2026 at 10:16 AM The Open Journals editorial robot <
***@***.***> wrote:

> *editorialbot* created an issue (openjournals/joss-reviews#10083)
> <https://github.com/openjournals/joss-reviews/issues/10083>
>
> *Submitting author:* @FrancescaZanderigo
> <https://github.com/FrancescaZanderigo> (Francesca Zanderigo
> <http://orcid.org/0000-0001-6510-0676>)
> *Repository:* https://github.com/ehauser-mind/PICNIC.git
> *Branch with paper.md* (empty if default branch): main
> *Version:* v0.1.4
> *Editor:* @mstimberg <https://github.com/mstimberg>
> *Reviewers:* @ghisvail <https://github.com/ghisvail>, @pjmark
> <https://github.com/pjmark>
> *Archive:* Pending
> Status
>
> [image: status]
> <https://joss.theoj.org/papers/e101744ccc49ec37a8dd74bcce1a0855>
>
> Status badge code:
>
>
> *Reviewers and authors*:
>
> Please avoid lengthy details of difficulties in the review thread.
> Instead, please create a new issue in the target repository
> <https://github.com/ehauser-mind/PICNIC.git> and link to those issues
> (especially acceptance-blockers) by leaving comments in the review thread
> below. (For completists: if the target issue tracker is also on GitHub,
> linking the review thread in the issue or vice versa will create
> corresponding breadcrumb trails in the link target.)
> Reviewer instructions & questions
>
> @ghisvail <https://github.com/ghisvail> & @pjmark
> <https://github.com/pjmark>, your review will be checklist based. Each of
> you will have a separate checklist that you should update when carrying out
> your review.
> First of all you need to run this command in a separate comment to create
> the checklist:
>
> @editorialbot generate my checklist
>
> The reviewer guidelines are available here:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any
> questions/concerns please let @mstimberg <https://github.com/mstimberg>
> know.
>
> ✨ *Please start on your review when you are able; the expectation is that
> it will take 2-4 weeks to complete.* ✨
> Checklists
>
> @ghisvail <https://github.com/ghisvail>, please create your checklist
> typing: @editorialbot generate my checklist
>
> @pjmark <https://github.com/pjmark>, please create your checklist typing: @editorialbot
> generate my checklist
>
