### 2026-06-24 1943372
Use greater than rather than approximately equal to allow more flexibility with libraries
