+++ requirements.txt
@@ -1,13 +1,13 @@
 # Included in yml file, must match versions
-numpy~=2.2.6
-scipy~=1.15.1
-matplotlib~=3.8
-pandas~=2.2
+numpy>=2.0
+scipy>=1.15
+matplotlib>=3.8
+pandas>=3.0
 # Independent of yml file
-setuptools~=80.9.0
+setuptools>=80.9.0
 FreeSimpleGUI~=5.2.0.post1
-pillow~=11.3.0
-nibabel~=5.3.2
-nilearn
-nipype~=1.10.0
-templateflow~=25.0.3
+pillow>=11.3.0
+nibabel>=5.3
+nilearn>=0.13
+nipype>=1.10.0
+templateflow>=25.0
