+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = picnic
-version = 0.1.7
+version = 0.1.8
 author = Eric Hauser
 author_email = eric.hauser@nyspi.columbia.edu
 maintainer = Mike Schmidt
