### 2026-09-02 #10083 @ghisvail
First of all, my sincere apologies for the delay. My review has been mostly ready for months, but could not find the time to finalize it lately. I sincerely thank all of you for your patience.

Second, I took this opportunity to check whether there has been updates to the paper or code that could have altered my initial impressions of the paper back then. It looks like neither received much activities so my comments below should still stand as of today.

# Review checklist from @ghisvail

## The paper

- [x] List of authors with affiliations
- [x] Summary of functionalities and purpose

Covered in paragraph 2 of the *Summary* section.

- [x] Statement of need

This part is clearly lacking. I would advise to spend less text describing the value and usage of the PET imaging modality, and more on explaining where the concrete value of the software stands.

-  [x] Comparison with state-of-the-art

This part is very weak too, since it's mixed within the statement of need. It is unclear as to me what features PICNIC offers over APPIAN (Funck et al. 2018) and Clinica (Routier et al. 2021). It's only covered by a single sentence, with vague wording which would suggest that both pieces of software are either closed-source, or lack modularity and flexibility in their design, which none applies in my opinion. I would advise to amend the sentence with more accurate and verifiable claims, such as the additional configurability options through the "input deck" text file, and the interactive QC features for instance. There could be more.

- [ ] (optional) Ongoing research projects using it
- [x] References

## Software license

The repository contains a license file with a copyright statement and the terms of the MIT license.

## Scope and significance

- [x] Published research citing or using the software

- [ ] Evidence of adoption by other research groups

No evidence, but probably not a deal breaker?

- [ ] Integration into established research workflows or pipelines

No evidence

- [ ] Demonstrated performance improvements or novel capabilities compared to alternatives.

No concrete evidence, only vague claims. Needs further work.

- [x] Clear potential for reuse and citation by other researchers in the domain

Authors have released the code on GitHub, and packaged the software with Docker. Availability could be further increased by releasing on PyPI too.

## Development timeline

- [x] Commit history

Spans roughly 2 years, 3 contributors (1 main), spread over by batches.

- [ ] Iterative development

Basically an initial commit (b2e6a71) followed by bug fixes and infrastructure setup.

- [ ] Evolution of features

No change log recording feature evolution
No release tags to compare versions
Not much evolution beyond the initial commit

## Open development

- [ ] Evidence of releases or version tags

Version bump commits at regular intervals.
No documented change log or release tags.

- [ ] Public issues and/or pull requests

A handful, with little to no description.

- [ ] External community engagement

No evidence in the issues / PRs.
No external committers.
Only a single fork.

## Collaborative effort

- [x] Multiple contributors to the codebase

- [ ] Code review through pull requests

No evidence of systematic code reviews.

- [ ] Responses to issues and feature requests

No evidence of systematic responses.

- [ ] Evidence of community-driven improvements

No evidence of sustained nor occasional community involvement.

## Good practices

- [x] OSI-approved license

Little information beyond the readme. There is a Sphinx documentation project, but it does not seem like it is published anywhere. From a potential user's POV, this is very much insufficient.

- [ ] Quality assurance

No CI, no tests, no documented verification process. Difficult to verify whether the software is working as intended.

- [ ] Releases

No tagged releases, changelog, or formal release process documented somewhere.

- [x] Community pathways

The project features a contributor's guide.

- [ ] Statement of need

Should be present in the readme.

Only Docker is covered. Outside Docker, the software is indeed pip installable, but no instructions are given for the external dependencies (Freesurfer, SPM, etc.), which is definitely not trivial.

- [x] Example usage

- [ ] API documentation

None

- [x] Community guidelines

Nothing to comment here, since the project does not have a test suite.

## Software paper content

- [x] Statement of need
Too superficial on details regarding novelty and filled gap.

- [ ] State of the field
Too superficial, missing the standout features justifying the development of a separate project.

- [ ] Software design
Design and architecture of the software are covered superficially.
Details regarding challenges, choices or reasoning are absent from the paper.

- [x] Research impact statement
There are multiple citations of publications from the same research group using the software.

- [ ] AI usage disclosure
No explicit disclosure on whether AI was used or not to help write the paper.
