### 2026-09-03 #10083 @mstimberg
Many thanks for your review @ghisvail, good to have you back I was afraid we lost you 😅
Could you please generate the default checklist with `@editorialbot generate my checklist`? No need to repeat the comments, just ticking the boxes would be enough – it's a formality, but that checklist is linked from the issue description and is often checked by the associate EiC before the final publication.

Just one comment with regard to your reviews @ghisvail and @pjmark : you based them on https://joss.readthedocs.io/en/latest/review_criteria.html and https://joss.readthedocs.io/en/latest/review_checklist.html, which would be perfect if this wasn't a submission from 2025, i.e. before JOSS changed its policy (see the [blog post](https://blog.joss.theoj.org/2026/01/preparing-joss-for-a-generative-ai-future)). For such a "pre-2026-submission", the relevant links are:
- https://joss.readthedocs.io/en/pre-2026-submission/review_checklist.html
- https://joss.readthedocs.io/en/pre-2026-submission/review_criteria.html

There is of course a big overlap in the two set of documents, so there is need to re-write/-structure your submitted reviews. Just two remarks about the main differences:
- The new [scope and significance](https://joss.readthedocs.io/en/latest/review_criteria.html#scope-and-significance) requirement is more demanding than the previous [substantial scholarly effort](https://joss.readthedocs.io/en/pre-2026-submission/review_criteria.html#substantial-scholarly-effort) criterion – the latter should be applied to this manuscript
- The [previous format](https://joss.readthedocs.io/en/pre-2026-submission/review_checklist.html#software-paper) that applies here only requires two mandatory sections in the paper: *Summary* and *Statement of Need*. The additional sections mentioned in the [newer guidelines](https://joss.readthedocs.io/en/latest/review_checklist.html#software-paper) (e.g. Software Design, AI usage disclosure) can of course be added by the authors if they prefer, but they are not mandatory.

With all that said in mind, please prepare your response and/or update the code/paper to address the reviewers' concerns, @FrancescaZanderigo. Please aim to do this in the time frame of 2–4 weeks or let me know if you think this is not feasible.

Of course (for all of you), please do not hesitate to contact me in case anything is unclear.
